Refactor DoctorService
Add UtilService with mapDoctorDtoToDoctor function, override DoctorDto.toString method and adapt DoctorServiceTest accordingly. Rename DoctorApiService
