+++ backend/src/main/java/de/neuefische/backend/service/UtilService.java
@@ -1,0 +1,25 @@
+package de.neuefische.backend.service;
+
+import de.neuefische.backend.dto.DoctorDto;
+import de.neuefische.backend.model.Doctor;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UtilService {
+
+    public Doctor mapDoctorDtoToDoctor(DoctorDto doctorDto) {
+            return Doctor.builder()
+                .firstName(doctorDto.getFirstName())
+                .lastName(doctorDto.getLastName())
+                .specialty(doctorDto.getSpecialty())
+                .street(doctorDto.getStreet())
+                .streetNumber(doctorDto.getStreetNumber())
+                .postalCode(doctorDto.getPostalCode())
+                .city(doctorDto.getCity())
+                .country(doctorDto.getCountry())
+                .phoneNumber(doctorDto.getPhoneNumber())
+                .mobileNumber(doctorDto.getMobileNumber())
+                .emailAddress(doctorDto.getEmailAddress())
+                .build();
+    }
+}
